feat: deal with the white list
