+++ constants/compensation_list.json
@@ -1,0 +1,6 @@
+{
+  "whitelist": [
+    "0x39A3cc8Da350b3c6B436Cb47FeAEE8aae54ACD28",
+    "0x1B1d9Cb1Ca240F7dD9359c33376c42f4c220C31c"
+  ]
+}
